Gets edit player/parent form created and pre-loaded with information about both.  TODO: fix the "Add to Roster" spinner to initialize with correct information.

+++ app/src/main/java/android/reserver/capstone_robertklare/DAO/ParentDAO.java
@@ -41,5 +41,5 @@ public interface ParentDAO {
     String getParentLastName(long id);
 
     @Query("DELETE FROM Parents WHERE personID = :id;")
-    void deleteParentByID(int id);
+    void deleteParentByID(long id);
 }
